+++ routes/subscription.js
@@ -20,18 +20,24 @@ function authenticateToken(req, res, next) {
 
 router.get("/", authenticateToken, async (req, res, next) => {
     try {
-        const allSubscriptions = await subscriptionDAO.getSubscriptions();
-        let result = [];
-        for (let i in allSubscriptions) {
-            let subscription = allSubscriptions[i]
-            const newSubscription = {
-                startDate: subscription.startDate.getFullYear() + "-" + (subscription.startDate.getMonth() + 1).toString().padStart(2, '0') + "-" + (subscription.startDate.getDate() + 1).toString().padStart(2, '0'),
-                name: subscription.name,
-                durationDays: subscription.durationDays
+        const userId = req.user._id;
+        const actualSubscription = await subscriptionDAO.getByUserId(userId);
+        if (!req.user.roles.includes('admin') && userId !== actualSubscription.userId) {
+            return res.status(401).send("Unauthorized action");
+        } else {
+            const allSubscriptions = await subscriptionDAO.getSubscriptions();
+            let result = [];
+            for (let i in allSubscriptions) {
+                let subscription = allSubscriptions[i]
+                const newSubscription = {
+                    startDate: subscription.startDate.getFullYear() + "-" + (subscription.startDate.getMonth() + 1).toString().padStart(2, '0') + "-" + (subscription.startDate.getDate() + 1).toString().padStart(2, '0'),
+                    name: subscription.name,
+                    durationDays: subscription.durationDays
+                }
+                result.push(newSubscription);
             }
-            result.push(newSubscription);
+            return res.json(result);
         }
-        return res.json(result);
     } catch (e) {
         console.log('Failed to get an item:', e)
         res.status(500).send(e.message);
@@ -58,8 +64,16 @@ router.post("/", authenticateToken, async (req, res, next) => {
 router.put("/:id", authenticateToken, async (req, res, next) => {
     try {
         const { durationDays, name, startDate } = req.body;
-        const updatedSubscription = await subscriptionDAO.updateById(req.params.id, { durationDays, name, startDate });
-        return res.json(updatedSubscription);
+        const userId = req.user._id;
+        const subscriptionId = req.params.id;
+        const actualSubscription = await subscriptionDAO.getById(subscriptionId);
+        if (!req.user.roles.includes('admin') && userId !== actualSubscription.userId) {
+            return res.status(401).send("Unauthorized action");
+        } else {
+            const updatedSubscription = await subscriptionDAO.updateById(req.params.id, { durationDays, name, startDate });
+            return res.json(updatedSubscription);
+        }
+
     } catch (e) {
         console.log(e)
         res.status(500).send(e.message);
@@ -69,9 +83,15 @@ router.put("/:id", authenticateToken, async (req, res, next) => {
 
 router.delete("/:id", authenticateToken, async (req, res, next) => {
     try {
+        const userId = req.user._id;
         const subscriptionId = req.params.id;
-        await subscriptionDAO.deleteById(subscriptionId);
-        return res.sendStatus(200);
+        const actualSubscription = await subscriptionDAO.getById(subscriptionId);
+        if (!req.user.roles.includes('admin') && userId !== actualSubscription.userId) {
+            return res.status(401).send("Unauthorized action");
+        } else {
+            await subscriptionDAO.deleteById(subscriptionId);
+            return res.sendStatus(200);
+        }
     } catch (e) {
         console.log(e)
         res.status(500).send(e.message);
